requirements.txt
# NEET Knowledge RAG Project
# A multi-format RAG system for NEET aspirants in India

# Core RAG Framework
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0

# Vector Database (FAISS)
faiss-cpu>=1.7.0

# Document Processing
pymupdf>=1.24.0  # PDF processing
pypdf>=4.0.0     # PDF reading
pytesseract>=0.3.10  # OCR for scanned PDFs
pdf2image>=1.17.0   # Convert PDF to images for OCR
pillow>=10.0.0   # Image processing

# YouTube & Video
youtube-transcript-api>=0.6.0  # YouTube transcripts
yt-dlp>=2024.0.0

# Web Content
beautifulsoup4>=4.12.0  # HTML parsing
requests>=2.31.0   # HTTP requests

# Text Processing
python-dotenv>=1.0.0  # Environment variables
tiktoken>=0.7.0    # Token counting

# LLM Providers
openai>=1.0.0   # OpenAI / OpenRouter API

# Quantized model inference (TranslateGemma 4-bit/8-bit)
bitsandbytes>=0.43.0
accelerate>=0.30.0

# Frontend
streamlit>=1.30.0

# Utilities
numpy>=1.24.0
pydantic>=2.0.0
pyyaml>=6.0
sentence-transformers>=2.2.0
redis>=4.0.0
boto3>=1.34.0
markdownify>=0.11.0
pandas>=2.0.0

# Type checking (dev)
mypy>=1.10.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0

# DB (async PostgreSQL for NEET Counselling bot)
asyncpg>=0.29.0

# Telegram Bot
python-telegram-bot>=21.0

# Hybrid Search
rank-bm25>=0.2.2

# Testing (async)
pytest-asyncio>=0.23.0
